refactor: layout components
