+++ src/components/FormSuccess.svelte
@@ -4,7 +4,7 @@
 	export let issue: number;
 	export let buttonWidth = 'w-52';
 
-	import HeaderPlaceholder from '$components/HeaderPlaceholder.svelte';
+	import HeaderPlaceholder from '$components/layout/HeaderPlaceholder.svelte';
 	import PrimaryButton from '$components/PrimaryButton.svelte';
 	import { theme } from '$lib/store';
 	import { detectTheme } from '$lib/utils';
