+++ src/components/OpenTicket.svelte
@@ -2,7 +2,7 @@
 	import Time from 'svelte-time';
 
 	import Icon from '$components/Icon.svelte';
-	import TicketLabel from './TicketLabel.svelte';
+	import TicketLabel from '$components/TicketLabel.svelte';
 
 	export let assignees: { html_url: string; avatar_url: string; login: string }[] = [];
 	export let comments: number = 0;
