+++ src/components/area/AreaMap.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import { browser } from '$app/environment';
 	import { GradeTable } from '$lib/constants';
-	import AreaMerchantDrawer from './AreaMerchantDrawer.svelte';
+	import AreaMerchantDrawer from '$components/area/AreaMerchantDrawer.svelte';
 	import Icon from '$components/Icon.svelte';
 	import MapLoadingEmbed from '$components/MapLoadingEmbed.svelte';
 	import ShowTags from '$components/ShowTags.svelte';
-	import TaggingIssues from '../TaggingIssues.svelte';
+	import TaggingIssues from '$components/TaggingIssues.svelte';
 	import { loadMapDependencies } from '$lib/map/imports';
 	import {
 		attribution,
