+++ src/components/area/AreaPage.svelte
@@ -6,11 +6,11 @@
 	export let type: 'country' | 'community';
 	export let data: AreaPageProps;
 
-	import AreaActivity from './AreaActivity.svelte';
-	import AreaMap from './AreaMap.svelte';
-	import AreaMerchantHighlights from './AreaMerchantHighlights.svelte';
-	import AreaStats from './AreaStats.svelte';
-	import AreaTickets from './AreaTickets.svelte';
+	import AreaActivity from '$components/area/AreaActivity.svelte';
+	import AreaMap from '$components/area/AreaMap.svelte';
+	import AreaMerchantHighlights from '$components/area/AreaMerchantHighlights.svelte';
+	import AreaStats from '$components/area/AreaStats.svelte';
+	import AreaTickets from '$components/area/AreaTickets.svelte';
 	import Boost from '$components/Boost.svelte';
 	import Icon from '$components/Icon.svelte';
 	import IssuesTable from '$components/IssuesTable.svelte';
