+++ src/components/layout/Footer.svelte
