+++ src/components/layout/HeaderPlaceholder.svelte
