+++ src/routes/tagger/[id]/+page.svelte
@@ -4,8 +4,8 @@
 
 	import { browser } from '$app/environment';
 	import { goto } from '$app/navigation';
-	import Footer from '$components/Footer.svelte';
-	import Header from '$components/Header.svelte';
+	import Footer from '$components/layout/Footer.svelte';
+	import Header from '$components/layout/Header.svelte';
 	import ProfileActivity from './components/ProfileActivity.svelte';
 	import ProfileStat from '$components/ProfileStat.svelte';
 	import Tip from '$components/Tip.svelte';
